delete references that are not cited
The original report lists a bunch of references, but doesn't actually
appear to use them in the text. For now, let's get rid of them. If it
turns out we'd like to include them in the bibliography even without
citing them, information on them will always be available in the
repository.