Update executive_summary.tex
I added line 19 "conclusion and future phases"  - because I think we will need some kind of closure.  We may want to state that we anticipate an R&D Phase II, or that there may be several of them anticipated as vendors pick up and begin to work with these technology recommendations.