Update e2e_viv_explained.tex
I only got through line 142 on this pass, but I am thinking about the Exec Summary and should really focus on that. I'm sorry that I have not offered more in this section. I know it needs to move to Lisa today. Some parts of this explanation section may work well in the Exec Summ which also needs to clearly state what E2E is, how it works/what it does, what at the benefits and what are the challenges we face.